fix: arc-weight/marking int→float truncation, transition_data recording, daemonic process error
- subnet_parameters_view.py: ListStore arc column int→float (0.1 no longer truncated to 0)
- viability_panel.py: place marking editor int()→float()
- batch_executor.py: removed skip-zeros logic in _apply_snapshot_to_model (CBD=0 controls work)
- batch_executor.py: _worker_run_experiment use_parallel=False + _SHYPN_IN_POOL_WORKER env var
- data_collector.py: transition_data always recorded in numpy fast-path
- batch_runner.py: total_firings dict added to both fork-based and sequential results
- sweep_output.py: fixed final_time NaN using time_points[-1]

Cross-validated: SP-Snapshot, SP-Batch, VP-Remote all match within 1% for 31 key places.
Glutathione: 515.7 (snapshot) vs 512.8 (batch) vs 510.5 (viability) — confirmed correct.

+++ src/shypn/ui/panels/viability/viability_panel.py
@@ -1161,7 +1161,7 @@ class ViabilityPanel(Gtk.Box):
     def _on_place_marking_edited(self, widget, path, new_text, store):
         """Handle place marking edit."""
         try:
-            new_marking = int(new_text)
+            new_marking = float(new_text)
             place_id = store[path][0]
             
             # Update store
@@ -1227,7 +1227,7 @@ class ViabilityPanel(Gtk.Box):
     def _on_arc_weight_edited(self, widget, path, new_text, store):
         """Handle arc weight edit."""
         try:
-            new_weight = int(new_text)
+            new_weight = float(new_text)
             arc_id = store[path][0]
             
             # Update store
